livekit/requirements.txt
livekit-agents[openai]
python-dotenv
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-elevenlabs
livekit-plugins-silero
livekit-plugins-google
boto3
psycopg2-binary
